requirements.txt
django==3.1
djangorestframework==3.11.0
mysqlclient==2.0.0
psycopg2==2.9.3
